+++ app/(home)/(tabs)/main/_layout.tsx
@@ -1,5 +1,5 @@
 import { Stack } from "expo-router";
 
-export default function ListLayout() {
+export default function MainLayout() {
   return <Stack screenOptions={{ headerShown: false }} />;
 }
