+++ app/(home)/(tabs)/main/index.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "expo-router";
 import { useContext } from "react";
 import { Pressable, Text, View } from "react-native";
-import { AuthContext } from "../_layout";
+import { AuthContext } from "../../_layout";
 
 export default function Main() {
   const { logout } = useContext(AuthContext);
@@ -16,30 +16,6 @@ export default function Main() {
       }}
     >
       <Text style={{ color: "black", fontSize: 16 }}>main</Text>
-
-      <Pressable
-        onPress={() => router.push("/(home)/mypage")}
-        style={{
-          backgroundColor: "#4A90E2",
-          paddingVertical: 12,
-          paddingHorizontal: 30,
-          borderRadius: 10,
-        }}
-      >
-        <Text style={{ color: "white", fontSize: 16 }}>마이페이지 이동</Text>
-      </Pressable>
-
-      <Pressable
-        onPress={() => router.push("/(home)/mypage")}
-        style={{
-          backgroundColor: "#4A90E2",
-          paddingVertical: 12,
-          paddingHorizontal: 30,
-          borderRadius: 10,
-        }}
-      >
-        <Text style={{ color: "white", fontSize: 16 }}>모아보기 이동</Text>
-      </Pressable>
 
       <Pressable
         onPress={logout}
