+++ app/(home)/(tabs)/mypage/_layout.tsx
@@ -1,5 +1,5 @@
 import { Stack } from "expo-router";
 
-export default function ListLayout() {
+export default function MypageLayout() {
   return <Stack screenOptions={{ headerShown: false }} />;
 }
